apps/llm-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python==0.2.27
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
